Master rework
Two different masters were created, one for dealing with static scheduling and one for dealing with event-driven dynamic scheduling, which needs feedback from slaves to proceed with the resource allocation.
