+++ include/ispd/message/message.hpp
@@ -25,10 +25,16 @@ struct ispd_message {
   int route_offset;
   tw_lpid previous_service_id;
 
+
+  /// \brief Sender information
+  tw_lpid service_id;
+
   /// \brief Message flags.
   unsigned int downward_direction: 1;
   unsigned int task_processed: 1;
-  unsigned int: 6; /// Reversed flags.
+  tw_lpid machine_id;
+  unsigned int: 5; /// Reversed flags.
 };
 
 #endif // ISPD_MESSAGE_H
+
