+++ include/ispd/scheduler/scheduler.hpp
@@ -25,7 +25,7 @@ public:
   /// This method is responsible for initializing any necessary data structures
   /// or state required by the scheduler before scheduling tasks.
   ///
-  virtual void initScheduler() = 0;
+  virtual void initScheduler(std::vector<tw_lpid> &slaves) = 0;
 
   /// \brief Performs forward scheduling of tasks.
   ///
